+++ src/Twilio/Rest/Api/V2010/Account/Sip/DomainContext.php
@@ -88,6 +88,8 @@ class DomainContext extends InstanceContext {
             'DomainName' => $options['domainName'],
             'EmergencyCallingEnabled' => Serialize::booleanToString($options['emergencyCallingEnabled']),
             'Secure' => Serialize::booleanToString($options['secure']),
+            'ByocTrunkSid' => $options['byocTrunkSid'],
+            'EmergencyCallerSid' => $options['emergencyCallerSid'],
         ]);
 
         $payload = $this->version->update('POST', $this->uri, [], $data);