+++ src/Twilio/Rest/Verify/V2/ServiceInstance.php
@@ -31,6 +31,7 @@ use Twilio\Version;
  * @property bool $dtmfInputRequired
  * @property string $ttsName
  * @property bool $doNotShareWarningEnabled
+ * @property bool $customCodeEnabled
  * @property \DateTime $dateCreated
  * @property \DateTime $dateUpdated
  * @property string $url
@@ -64,6 +65,7 @@ class ServiceInstance extends InstanceResource {
             'dtmfInputRequired' => Values::array_get($payload, 'dtmf_input_required'),
             'ttsName' => Values::array_get($payload, 'tts_name'),
             'doNotShareWarningEnabled' => Values::array_get($payload, 'do_not_share_warning_enabled'),
+            'customCodeEnabled' => Values::array_get($payload, 'custom_code_enabled'),
             'dateCreated' => Deserialize::dateTime(Values::array_get($payload, 'date_created')),
             'dateUpdated' => Deserialize::dateTime(Values::array_get($payload, 'date_updated')),
             'url' => Values::array_get($payload, 'url'),