+++ tests/Twilio/Integration/Supersim/V1/FleetTest.php
@@ -44,6 +44,7 @@ class FleetTest extends HolodeckTestCase {
                 "commands_enabled": true,
                 "commands_method": "GET",
                 "commands_url": "https://google.com",
+                "network_access_profile_sid": "HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "url": "https://supersim.twilio.com/v1/Fleets/HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
@@ -83,6 +84,7 @@ class FleetTest extends HolodeckTestCase {
                 "commands_enabled": true,
                 "commands_method": "POST",
                 "commands_url": null,
+                "network_access_profile_sid": "HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "url": "https://supersim.twilio.com/v1/Fleets/HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
@@ -115,13 +117,13 @@ class FleetTest extends HolodeckTestCase {
             {
                 "fleets": [],
                 "meta": {
-                    "first_page_url": "https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0",
+                    "first_page_url": "https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0",
                     "key": "fleets",
                     "next_page_url": null,
                     "page": 0,
                     "page_size": 50,
                     "previous_page_url": null,
-                    "url": "https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0"
+                    "url": "https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0"
                 }
             }
             '
@@ -138,13 +140,13 @@ class FleetTest extends HolodeckTestCase {
             '
             {
                 "meta": {
-                    "first_page_url": "https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0",
+                    "first_page_url": "https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0",
                     "key": "fleets",
                     "next_page_url": null,
                     "page": 0,
                     "page_size": 50,
                     "previous_page_url": null,
-                    "url": "https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0"
+                    "url": "https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0"
                 },
                 "fleets": [
                     {
@@ -157,6 +159,7 @@ class FleetTest extends HolodeckTestCase {
                         "commands_enabled": true,
                         "commands_method": "POST",
                         "commands_url": null,
+                        "network_access_profile_sid": "HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "sid": "HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "url": "https://supersim.twilio.com/v1/Fleets/HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
@@ -198,6 +201,7 @@ class FleetTest extends HolodeckTestCase {
                 "commands_enabled": true,
                 "commands_method": "POST",
                 "commands_url": null,
+                "network_access_profile_sid": "HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "url": "https://supersim.twilio.com/v1/Fleets/HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }