+++ CHANGES.md
@@ -1,6 +1,45 @@
 twilio-php Changelog
 ====================
 
+[2020-05-20] Version 6.5.0
+--------------------------
+**Library - Feature**
+- [PR #635](https://github.com/twilio/twilio-php/pull/635): add __isset magic method for resource instance properties. Thanks to [@eshanholtz](https://github.com/eshanholtz)!
+- [PR #633](https://github.com/twilio/twilio-php/pull/633): add regional and edge support. Thanks to [@eshanholtz](https://github.com/eshanholtz)!
+
+**Library - Fix**
+- [PR #634](https://github.com/twilio/twilio-php/pull/634): env var retrieval and url unparsing. Thanks to [@eshanholtz](https://github.com/eshanholtz)!
+
+**Api**
+- Add optional `emergency_caller_sid` parameter to SIP Domain
+- Updated `call_reason` optional property to be treated as PII
+- Added optional BYOC Trunk Sid property to Sip Domain API resource
+
+**Autopilot**
+- Add Restore resource to Autopilot Assistant
+
+**Contacts**
+- Added contacts Create API definition
+
+**Events**
+- Subscriptions API initial release
+
+**Numbers**
+- Add Evaluations API
+
+**Supersim**
+- Allow filtering the Fleets resource by Network Access Profile
+- Allow assigning a Network Access Profile when creating and updating a Fleet
+- Add Network Access Profiles resource
+
+**Verify**
+- Add `CustomCode` optional parameter on Verification creation.
+- Add delete action on Service resource.
+
+**Voice**
+- Added endpoints for BYOC trunks, SIP connection policies and source IP mappings
+
+
 [2020-04-29] Version 6.4.0
 --------------------------
 **Library - Feature**